Rename: .ts파일 types폴더로 옮김 Feat: getServerSideProps로 로그인했을때 redirect 설정
