+++ components/header/HeaderLoginButton.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import styled, { css } from 'styled-components';
-import { StyleProps } from './header-props';
+import { StyleProps } from '../../types/header-props';
 
 const HeaderLoginButton = (props: StyleProps) => {
   return (
