+++ components/signup/SignupName.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import useInput from '../../hooks/use-input';
 import Input from '../UI/Input';
-import { checkProps } from './signup-props';
+import { checkProps } from '../../types/signup-props';
 import { ErrorMessage, SignupInput, SignupInputText } from './SignupForm';
 
 const nameRegex = /^[가-힣]{2,5}$/;
