+++ types/input-props.ts
