+++ types/ridi-cash-props.ts
