+++ types/style.d.ts
